# Service worker cache for faster local frontend development (SCP-4379)