--- a/app/javascript/lib/metrics-perf.js
+++ b/app/javascript/lib/metrics-perf.js
@@ -214,7 +214,6 @@ export function calculatePerfTimes(perfTimes) {
     perfProps = Object.assign({
       'perfTime:cache': 'service-worker',
       'perfTime:serviceWorkerCacheHit': true,
-      'perfTime:serviceWorkerCacheEnabled': perfTimes.serviceWorkerCacheEnabled,
       'perfTime:url': perfTimes.url
     }, perfProps)
     return perfProps