--- a/app/javascript/lib/scp-api.jsx
+++ b/app/javascript/lib/scp-api.jsx
@@ -777,6 +777,22 @@ export function getFullUrl(path, mock=false) {
   return fullPath
 }
 
+/** Fetch, leveraging web cache if available */
+async function fetchWithWebCache(url, init) {
+  const webCache = await caches.open(`scp-${env}-1.6.0`)
+  let response = await webCache.match(url)
+  let hitOrMiss = 'hit'
+  if (typeof response === 'undefined') {
+    console.log(`Cache miss for SCP API fetch of URL: ${url}`)
+    response = await fetch(url, init).catch(error => error)
+    await webCache.put(url, response)
+    hitOrMiss = 'miss'
+    return await webCache.match(url)
+  }
+  console.log(`Cache ${hitOrMiss} for SCP API fetch of URL: ${url}`)
+  return await webCache.match(url)
+}
+
 /**
  * Client for SCP REST API.  Less fetch boilerplate, easier mocks.
  *
@@ -791,7 +807,8 @@ export default async function scpApi(
 
   const perfTimeStart = performance.now()
 
-  const response = await fetch(url, init).catch(error => error)
+  // const response = await fetch(url, init).catch(error => error)
+  const response = await fetchWithWebCache(url, init)
 
   // Milliseconds taken to fetch data from API
   // Suboptimal, but retained until at least Q4 2021 for continuity.