… app/javascript/lib/service-worker-cache.js
… app/javascript/lib/validation/validate-file-content.js
… app/javascript/providers/SCPContextProvider.jsx
… app/javascript/vite/application.js
… package.json
… test/js/lib/scp-api.test.js
… test/js/setup-tests.js
… version.txt
… vite.config.ts
… yarn.lock